==> diplom/1.ipynb <==
## «Модель прогнозирования стоимости жилья для агентства недвижимости»



# Дипломный проект

Data Science

Описание данных:

'status' — статус продажи;
'private pool' и 'PrivatePool' — наличие собственного бассейна;
'propertyType' — тип объекта недвижимости;
'street' — адрес объекта;
'baths' — количество ванных комнат;
'homeFacts' — сведения о строительстве объекта (содержит несколько
типов сведений, влияющих на оценку объекта);

'fireplace' — наличие камина;
'city' — город;
'schools' — сведения о школах в районе;
'sqft' — площадь в футах;
'zipcode' — почтовый индекс;
'beds' — количество спален;
'state' — штат;
'stories' — количество этажей;
'mls-id' и 'MlsId' — идентификатор MLS (Multiple Listing Service, система
мультилистинга);

target' — цена объекта недвижимости (целевой признак, который
необходимо спрогнозировать).

import numpy as np
import pandas as pd
import re
from geopy.geocoders import Nominatim
from fake_useragent import UserAgent
from geopy.exc import GeocoderTimedOut, GeocoderUnavailable
from geopy.exc import GeocoderServiceError

Предварительный просмотр

df = pd.read_csv('data/data.csv')
display(df.head())
df.info()

Пропуски

missing_data = df.isnull().sum()

for column_name, missing_count in missing_data.items():
    if missing_count > 0:
        percentage = (missing_count / len(df)) * 100
        print(f'{column_name}: {missing_count} пропусков ({percentage:.2f}%)')

Явные дубликаты


rows_before = df.shape[0] #удалим дубликаты
df = df.drop_duplicates(ignore_index=True)
rows_after = df.shape[0]
duplicates_removed = rows_before - rows_after
print(f'Удалено {duplicates_removed} дубликатов')

Дополнительные дубликты

print(df['private pool'].unique())
print(df['PrivatePool'].unique())


df['pool_encoded'] = df['private pool'].combine_first(df['PrivatePool'])
df = df.drop(['PrivatePool', 'private pool', 'MlsId', 'mls-id'], axis=1)
df['pool_encoded'] = df['pool_encoded'].fillna(False)
df['pool_encoded'] = df['pool_encoded'].replace(['yes', 'Yes'], True)
rows_before = df.shape[0] #удалим дубликаты
df = df.drop_duplicates(ignore_index=True)
rows_after = df.shape[0]
duplicates_removed = rows_before - rows_after
print(f'Удалено {duplicates_removed} дубликатов')

Признак homeFacts

missing_values_count = df.homeFacts.isna().sum()
total_count = len(df)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df.homeFacts.nunique()}\n"
f"Уникальные значения: {list(df.homeFacts.unique()[:1])}")

label = df.homeFacts.str.findall(r"\bfactLabel': ([\s\S]+?)[}\b]")
label[100]

Value = df.homeFacts.str.findall(r"\bfactValue': ([\s\S]+?), 'factLabel\b")
Value[:10]

list_label = ','.join(label[0]).replace("'","").split(',')
list_label

for i, val in enumerate(list_label):
    df[val]=Value.apply(lambda x: x[i])

# удаляем признак homeFacts 
df = df.drop('homeFacts', axis=1)

 Признак propertyType

missing_values_count = df.propertyType.isna().sum()
total_count = len(df)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
      f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
      f"Уникальные значения: {df.propertyType.nunique()}\n"
      f"Уникальные значения: {list(df.propertyType.unique()[:10])}")

df.propertyType = df.propertyType.str.lower()
missing_values_count = df.propertyType.isna().sum()
total_count = len(df)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
      f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
      f"Уникальные значения: {df.propertyType.nunique()}\n"
      f"Уникальные значения: {list(df.propertyType.unique()[:10])}")

df.propertyType = df.propertyType \
.str.replace('single-family home','single family') \
.str.replace('single family home','single family') \
.str.replace('yes','Other') \
.str.replace('unknown','Other')
df['Type'] = df['propertyType'].str.split(',').str[0]

missing_values_count = df.Type.isna().sum()
total_count = len(df)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
      f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
      f"Уникальные значения: {df.Type.nunique()}\n"
      f"Уникальные значения: {list(df.Type.unique()[:10])}")

synonyms_mapping = {
    'single_family_home': [
        'single family', '1 story', '2 story', 'detached', '1 story/ranch', '1 story traditional', 'single detached', 'singlefamilyresidence', 'single wide', 'single-wide mobile with land',
        'two story', 'one story', 'one level unit', 'rancher', '1 1/2 story', 'single wide mh', 'one story traditional'],
    'multi_family_home': [
        'multi-family', 'multi-family home', 'duplex', 'triplex', 'fourplex', 'multi_level', 'multi family', '2 unit condo', '2-story'],
    'condo': [
        'condo', 'coop', 'cooperative', 'condo/townhome/row home/co-op', 'condo/townhome', 'condominium', 'condo/unit', 'apartment/condo/townhouse', 'co-op', '2 story condo', 'high rise'],
    'townhouse': [
        'townhouse', 'townhome style', 'townhouse-interior', 'townhouse-end unit'],
    'apartment': [
        'apartment', 'condominium (single level)', 'high-rise', 'mid-rise', 'low-rise (1-3 stories)', 'Flats', 'studio'],
    'land': [
        'lot/land', 'land'],
    'mobile_home': [
        'mobile/manufactured', 'mobile / manufactured', 'manufactured house', 'mfd/mobile home', 'mobile home', 'manufactured home', 'manufactured double-wide', 'manufactured single-wide', 'mobile home 1 story', 'mobile manu - double wide'],
    'miscellaneous': [
        'miscellaneous'],
    'ranch': [
        'ranch', 'one story'],
    'modern': [
        'contemporary', 'contemporary/modern', "modern", 'mid-century modern', 'modern farmhouse', 'modernist'],
    'historical': [
        'historical', 'designated historical home', 'historical/conservation district', 'historic/older', 'historic vintage', 'historic'],
    'other': [
        'Other', 'english', 'urban contemporary', 'other style', 'florida', 'farms/ranches', 'carriage house', 'country english', 'straight thru', 'less than 4 floors', 'bungalow', 'custom', 'arts & crafts', 'tudor', 'new build 2019', 'split foyer', 'cottage', 'cottage/camp', 'garden home', 'farm/ranch', 'farm/ranch house', 'farm house', 'hi ranch', 'attached duplex', 'farmhouse', 'houseboat', 'ground floor', 'victorian', '3 story', '3+ stories', 'santa barbara/tuscan', 'old style', 'modular/prefab', 'post and beam', 'manuf/mobile', 'multiple occupancy', 'attached', 'hawaiian plantation', 'forest garden home', '1 1/2 story with basement', 'split-entry', 'texas hill country', 'lake house', '1 story with basement', 'hi-rise', 'coastal beach home', 'historical', 'key west/coastal', 'loft/balcony', 'english manor', 'mid-rise (4-7 stories)', 'mid-level', 'new englander', 'residential (<1 acre)', 'ranch','residential (1+ acre)', 'split', 'split level', 'split (4 level)', 'split (5+ level)', 'urban', 'patio', 'patio home', 'penthouse', 'manor', 'victorian/federal', 'coastal', 'coastal contemporary', 'coastal ii', 'coastal modern', 'coastal two story', 'mountain contemporary', 'key west', 'high ranch', 'end unit']
}


def replace_synonyms(value):
    for key, synonym_values in synonyms_mapping.items():
        if value in synonym_values:
            return key
    return "other"

df['Type'] = df['Type'].apply(replace_synonyms)
missing_values_count = df.Type.isna().sum()
total_count = len(df)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
      f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
      f"Уникальные значения: {df.Type.nunique()}\n"
      f"Уникальные значения: {list(df.Type.unique()[:10])}")

# удаляем propertyType 
df.drop('propertyType', axis=1, inplace=True)

df1=df.copy()

Признак Remodeled year

df1['Remodeled year'] = df1['Remodeled year'].str.replace("'",'', regex=True)
mask= df1['Remodeled year'].isin(['', '0', '1111'])
df1.loc[mask,'Remodeled year'] = 'None' #уникальные значения
display(df1['Remodeled year'].sort_values().unique()[:10])

missing_values_count = df1['Remodeled year'].isna().sum()
total_count = len(df1)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df1['Remodeled year'].nunique()}\n"
f"Уникальные значения: {list(df1['Remodeled year'].sort_values().unique()[:10])}")

#None
print(f"Процент None: {round(df1[df1['Remodeled year']=='None'].shape[0]/df.shape[0]*100, 2)}%")
print(f"Количество None: {df1.loc[df1['Remodeled year']=='None', 'Remodeled year'].count()}")

# удаляем 'Remodeled year'
df1 = df1.drop('Remodeled year', axis=1)

признак Cooling

df1['Cooling'].str.lower().value_counts().head(10)

df1['Cooling'] = df1['Cooling'].str.replace("'",'', regex=True)
df1['Cooling_encoded'] = df1['Cooling'].apply(lambda x: True if x not in ['', 'no data', 'None', 'none'] else False)

# удаляем признак 'Cooling'
df1= df1.drop('Cooling', axis=1)

Признак Heating

display(df1['Heating'].str.lower().value_counts().head(10))

df1['Heating'] = df1['Heating'].str.replace("'",'', regex=True)
df1['Heating_encoded'] = df1['Heating'].apply(lambda x: True if x not in ['', 'no data', 'None', 'none'] else False)

missing_values_count = df1['Heating_encoded'].isna().sum()
total_count = len(df1)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения:: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df1['Heating_encoded'].nunique()}\n"
f"Уникальные значения: {list(df1['Heating_encoded'].sort_values().unique())}")

# удаляем признак 'Heating'
df1= df1.drop('Heating', axis=1)

Признак city

df1['city'] = df1['city'].str.title()
missing_values_count = df1['city'].isna().sum()
total_count = len(df1)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df1['city'].nunique()}\n"
f"Уникальные значения: {list(df1['city'].sort_values().unique()[:10])}")

df1['city'].str.lower().value_counts().sort_index().head(10)


df1['city'] = df1['city'].str.replace(' City', '') 
initial_row_count = df1.shape[0]
df1 = df1[~df1['city'].isin([' ', '--', np.nan])]
removed_row_count = initial_row_count - df1.shape[0]
print(f"Удаленные строки: {removed_row_count}")


missing_values_count = df1['city'].isna().sum()
total_count = len(df1)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df1['city'].nunique()}\n"
f"Уникальные значения: {list(df1['city'].sort_values().unique()[:10])}")

Признак sqft

missing_values_count = df1['sqft'].isna().sum()
total_count = len(df1)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df1['sqft'].nunique()}\n"
f"Уникальные значения: {list(df1['sqft'].sort_values().unique()[:10])}")

s_sqft = df1.sqft[df1.sqft.notna()]
s_sqft = s_sqft.str.findall(r"\d+")
s_sqft = s_sqft.apply(lambda x: ''.join(x))
df1.loc[df1.sqft.notna(), 'sqft']=s_sqft
df1.sqft = df1.sqft.fillna(0)
df1.loc[df1.sqft=='','sqft']=0
df1.sqft = df1.sqft.astype(int)

selected_rows = df1[(df1['sqft'] == 0) & ((df1['Type'] == 'other') | (df1['Type'] == 'land') | (df1['Type'] == 'ranch'))]
percent_selected = len(selected_rows) / len(df1) * 100
# Выводим полученные строки и процент выбранных строк
print(f"Выбранные строки: {len(selected_rows)}")
print(f"Выбранные строки: {percent_selected:.2f}%")

selected_rows = df1[(df1['sqft'] == 0) & ~(df1['Type'].isin(['other', 'land', 'ranch']))]
unique_statuses = selected_rows['Type'].unique()
# Выводим полученные строки и процент выбранных строк
percent_selected = len(selected_rows) / len(df1) * 100
print(f"Выбранные строки: {len(selected_rows)}")
print(f"Выбранные строки: {percent_selected:.2f}%")

Признак baths

missing_values_count = df1.baths.isna().sum()
total_count = len(df1)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
      f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
      f"Уникальные значения: {df1.baths.nunique()}\n"
      f"Уникальные значения: {list(df1.baths.unique()[:10])}")

df1.baths = df1.baths.str.replace('[a-zA-Z:]','', regex=True)
df1.baths = df1.baths.str.replace(' ','', regex=True)

df1['baths'] = df1['baths'].str.replace('[a-zA-Z+:]','', regex=True)
df1['baths'].replace(['nan','', '~', '..', '--', '—','0.0','0.00','0/0'], 0, inplace=True)
df1['baths'].replace(['1-2','1/1-0/1-0/1-0', '1/1/1/1','1-0/1-0/1', ], 1, inplace=True)
df1['baths'].replace(['2-1/2-1/1-1/1-1','3-1/2-2',], 2, inplace=True)
df1['baths'].replace(['116/116/116'], 116, inplace=True)
df1['baths'].replace(['7,500'], 7.5, inplace=True)
df1['baths'].replace(['5,000'], 5, inplace=True)
df1['baths'].replace(['3,500'], 3.5, inplace=True)
df1['baths'].replace(['2,750'], 2.75, inplace=True)
df1['baths'].replace(['2,250'], 2.25, inplace=True)
df1['baths'].replace(['1,250'], 1.25, inplace=True)
df1['baths'].replace(['1,500'], 1.5, inplace=True)
df1['baths'].replace(['4,000'], 4, inplace=True)
df1['baths'].replace(['2,000'], 2, inplace=True)
df1['baths'].replace(['3,000'], 3, inplace=True)
df1['baths'].replace(['2,500'], 2.5, inplace=True)
df1['baths'].replace(['1,000'], 1, inplace=True)
df1['baths'].replace(['1,750'], 1.75, inplace=True)
df1['baths'] = df1['baths'].fillna(0)

df1['baths'] = df1['baths'].astype(float).round()
selected_rows = df1[(df1['baths'] == 0) & ((df1['Type'] == 'other') | (df1['Type'] == 'land') | (df1['Type'] == 'ranch'))]
display(selected_rows.head())
percent_selected = len(selected_rows) / len(df1) * 100

print(f"Выбранные строки: {percent_selected:.2f}%")

selected_rows = df1[(df1['baths'] == 0) & ~(df1['Type'].isin(['other', 'land', 'ranch']))]
unique_statuses = selected_rows['Type'].unique()
print(f"Уникальные значения признака Type: {unique_statuses}")
percent_selected = len(selected_rows) / len(df1) * 100
print(f"Выбранные строки: {percent_selected:.2f}%")


missing_values_count = df1.baths.isna().sum()
total_count = len(df1)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
      f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
      f"Уникальные значения: {df1.baths.nunique()}\n"
      f"Уникальные значения: {list(df1.baths.unique()[:10])}")

Признак zipcode

df2=df1.copy()

missing_values_count = df2['zipcode'].isna().sum()
total_count = len(df2)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df2['zipcode'].nunique()}\n"
f"Уникальные значения: {list(df2['zipcode'].sort_values().unique()[:10])}")

df2[(df2['zipcode']=='--') | (df2['zipcode']=='0')| (df2['zipcode']=='00000') ]

indexes_to_drop = df2[(df2['zipcode']=='--') | (df2['zipcode']=='0')| (df2['zipcode']=='00000')].index
df2 = df2.drop(indexes_to_drop)
df2.zipcode.sort_values().unique()[:10]

df2.zipcode = df2.zipcode.str.replace("-.+",'', regex=True)
df2.zipcode.sort_values().unique()[:10]

Признак beds

missing_values_count = df2['beds'].isna().sum()
total_count = len(df2)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df2['beds'].nunique()}\n"
f"Уникальные значения: {list(df2['beds'].sort_values().unique()[:10])}")

# удаляем признак Beds
df2 = df2.drop('beds', axis=1)

Признак stories

missing_values_count = df2['stories'].isna().sum()
total_count = len(df2)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df2['stories'].nunique()}\n"
f"Уникальные значения: {list(df2['stories'].sort_values().unique()[:10])}")

df2[df2.stories.isna()].Type.value_counts()

# удаляем признак Beds
df2 = df2.drop('stories', axis=1)

Признак Target

missing_values_count = df2['target'].isna().sum()
total_count = len(df2)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
f"Уникальные значения: {df2['target'].nunique()}\n"
f"Уникальные значения: {list(df2['target'].sort_values().unique()[:10])}")

df2 = df2.dropna(subset=['target'])
df2.loc[:,'target'] = df2.loc[:,'target'].str.replace('[^0-9]','',regex=True)
df2.target = df2.target.astype(int)
display(df2[df2['target'] == df2['target']. min ()].head())
display(df2[df2['target'] == df2['target']. max ()].head())

count = len(df2[df2['target'] < 900])
percentage = count / len(df2) * 100
print(f"Количество строк: {count}")
print(f"Процент строк: {percentage:.2f}%")

df2 = df2.drop(df2[df2['target'] < 900].index)

display(df2[df2['target'] == df2['target']. min ()].head())

missing_data = df2.isnull().sum()
missing_columns = missing_data[missing_data > 0]

if missing_columns.empty:
    print('Пропуски отсутствуют')
else:
    for column_name in missing_columns.index:
        missing_count = missing_columns[column_name]
        percentage = (missing_count / len(df2)) * 100
        print(f'{column_name}: {missing_count} Пропуски ({percentage:.2f}%)')

признак Fireplace

df2['fireplace'].str.lower().value_counts().head(10)

df2['fireplace_encoded'] = df1['fireplace'].apply(lambda x: True if x not in ['', 'no data', 'None', 'none', '0', 'not applicable', 'no'] and not pd.isna(x) else False)


# удаляем признак fireplace
df2.drop('fireplace', axis=1, inplace=True)

Признак Status

missing_values_count = df2.status.isna().sum()
total_count = len(df2)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Количество пропущенных значений: {missing_values_count}\n"
      f"Процент пропущенных значений: {missing_values_percentage:.2f}%\n"
      f"Количество уникальных значений: {df2.status.nunique()}\n"
      f"Уникальные значения: {list(df2.status.unique()[:10])}")

categories = {
    "For Sale": ["for sale", "For sale", "New construction", "New"],
    "Active": [
        "Active", "A Active", "Active/Contingent", "Active Under Contract", "Active Option", "Auction - Active",
        "Active With Contingencies", "Active Option Contract", "Active Contingency", "Active Backup",
        "Active Contingent", "Active - Auction", "Active With Offer", "Active - Contingent", "Active with Contract",
        "Temporary Active", "Re Activated", "Reactivated"
    ],
    "Pending": [
        "P", "Pending", "pending", "P Pending Sale", "Pending Ab", "Pending Continue To Show",
        "Pending Inspection", "Pending Offer Approval", "Pending In", "Pending W/Insp Finance", "Pending Fe",
        "Pending W/Backup Wanted", "Pending Backups Wanted", "Pending With Contingencies", "Lease/Purchase Pending",
        "Pending Bring Backup", "Pending - Taking Backups", "Pending - Continue to Show",
        "Pending Taking Backups", "Offer Pending Signature", "Pending (Do Not Show)", "Pending W/ Cont.",
        "Pending W/Escape Clause", "Pending - Backup Offer Requested", "Pending Sale"
    ],
    "Contingent": [
        "Contingent", "Contingent Finance And Inspection", "Contingent Show",
        "Contingent Take Backup", "Contingent - Sale of Home", "Contingent Finance and Inspection",
        "C Continue Show", "Contingent   Show", "Contingent   Release", "Contingent   No Show",
        "CT Insp - Inspection Contingency", "Contingent   Foreclosure", "Conting Accpt Backups",
        "Contingent - Financing", "Contingency 48 Hr (+/ )", "Contingency Contract", "Contingent Escape"
    ],
    "Foreclosure": [
        "Pre-foreclosure", "Pre-foreclosure / auction", " / auction", "Foreclosed", "foreclosure", "Foreclosure"
    ],
    "Under Contract": [
        "Under Contract", "Under Contract   Showing", "Under Contract Backups", "Under Contract Show",
        "Under Contract - Show", "Under Contract - No Show", "Under contract", "U Under Contract",
        "Due Diligence Period", "Contract Contingent On Buyer Sale", "Contract P", "Ct", "Uc Continue To Show",
        "Under Contract Taking Back Up Offers", "Under Contract W/ Bckp", "Contract"
    ],
    "For Rent": ["for rent", "Apartment for rent", "Condo for rent"],
    "Auction": ["Auction"],
    "Recently Sold": ["recently sold"],
    "Price Change": ["Price Change"],
    "Back on Market": ["Back on Market", "Back On Market"],
    "Closed": ["Closed"],
    "Listing Extended": ["Listing Extended"],
    "Coming Soon": [
        "Coming soon: Nov 21.", "Coming soon: Dec 4.", "Coming soon: Nov 23.", "Coming soon: Nov 29.",
        "Coming soon: Dec 2.", "Coming soon: Dec 10.", "Coming soon: Dec 24.", "Coming soon: Nov 14.",
        "Coming soon: Nov 22.", "Coming soon: Oct 21.", "Coming soon: Dec 14.", "Coming soon: Oct 24.",
        "Coming soon: Dec 18.", "Coming soon: Dec 16.", "Coming soon: Dec 3.", "Coming soon: Dec 25.",
        "Coming soon: Nov 11.", "Coming soon: Nov 28.", "Coming soon: Nov 17.", "Coming soon: Dec 6.",
        "Coming soon: Nov 27.", "Coming soon: Nov 26.", "Coming soon: Dec 7.", "Coming soon: Dec 27.",
        "Coming soon: Dec 11.", "Coming soon: Dec 5.", "Coming soon: Nov 13.", "Coming soon: Nov 19.",
        "Coming soon: Nov 8.", "Coming soon: Oct 29.", "Coming soon: Dec 15.", "Coming soon: Oct 30.",
        "Coming soon: Dec 9.", "Coming soon: Dec 20.", "Coming soon: Dec 13.", "Coming soon: Dec 23.",
        "Coming soon: Nov 30.", "Coming soon: Dec 1.", "Coming soon: Nov 5.", "Coming soon: Nov 12.",
        "Coming soon: Nov 25.", "Coming soon: Nov 9."
    ],
}

def categorize_status(status_value):
    for category, values in categories.items():
        if status_value in values:
            return category
    return "Other"

# заменим значения на категории
df2['status'] = df2['status'].apply(categorize_status)

# выполним проверку
missing_values_count = df2.status.isna().sum()
total_count = len(df2)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
      f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
      f"Уникальные значения: {df2.status.nunique()}\n"
      f"Уникальные значения: {list(df2.status.unique()[:10])}")




rows_before = df2.shape[0]
df2= df2.drop_duplicates(ignore_index=True)
rows_after = df2.shape[0]
duplicates_removed = rows_before - rows_after
print(f'Удалено {duplicates_removed} дубликатов')

Признак  Street

missing_values_count = df2.street.isna().sum()
total_count = len(df)
missing_values_percentage = (missing_values_count / total_count) * 100

print(f"Пропущенные значения: {missing_values_count}\n"
      f"Пропущенные значения: {missing_values_percentage:.2f}%\n"
      f"Уникальные значения: {df.street.nunique()}")

# удаляем строки, где есть пропуски в столбце street
df2 = df2.dropna(subset=['street'])

df3=df2.copy()

missing_data = df3.isnull().sum()
missing_columns = missing_data[missing_data > 0]

if missing_columns.empty:
    print('Пропуски отсутствуют')
else:
    for column_name in missing_columns.index:
        missing_count = missing_columns[column_name]
        percentage = (missing_count / len(df2)) * 100
        print(f'{column_name}: {missing_count} Пропуски ({percentage:.2f}%)')

df3.duplicated().sum() 

# Сохраняеи данные в CSV-файл
df3.to_csv("data/cleaned1_data.csv", index=False)
